+++ packages/components-js/scripts/convertToReactVRTPage.ts
@@ -45,8 +45,8 @@ export const convertToReactVRTPage = (
     pdsImports && `import { ${pdsImports} } from '@porsche-design-system/components-react';`,
     reactImports && `import { ${reactImports} } from 'react';`,
     isIconPage && `import { ICON_NAMES } from '@porsche-design-system/icons';`,
-    (usesSetAllReady || usesComponentsReady) && `import { pollComponentsReady } from '../pollComponentsReady';`,
-    usesToast && `import { Toast } from '../components';`,
+    (usesSetAllReady || usesComponentsReady) && `import { pollComponentsReady } from '../../pollComponentsReady';`,
+    usesToast && `import { Toast } from '../../components';`,
   ]
     .filter(Boolean)
     .join('\n');