+++ packages/components/CHANGELOG.md
@@ -25,6 +25,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   ([#2676](https://github.com/porsche-design-system/porsche-design-system/pull/2676))
 - Vue: `cdn: 'auto' | 'cn'` prop to `PorscheDesignSystemProvider` as alternative to using
   `window.PORSCHE_DESIGN_SYSTEM_CDN` ([#2676](https://github.com/porsche-design-system/porsche-design-system/pull/2676))
+- Support for sticky footer to `Modal`
+  ([#2723](https://github.com/porsche-design-system/porsche-design-system/pull/2723))
 
 #### Changed